解决generated目录下子目录code被删除问题

+++ Setup/Console/Setup/Upgrade.php
@@ -73,9 +73,17 @@ class Upgrade implements \Weline\Framework\Console\CommandInterface
             }
         }
         $i += 1;
-        $this->printer->warning($i . '、generated生成目录代码code清理...', '系统');
-        $this->system->exec('rm -rf ' . Env::path_framework_generated_code);
-        $this->system->exec('mkdir  ' . Env::path_framework_generated_code);
+        if(!is_dir(Env::path_framework_generated_code)){
+            mkdir(Env::path_framework_generated_code,755,true);
+        }
+        // 扫描目录子目录
+        $this->printer->note($i. '、目录扫描...', '系统');
+        $this->printer->note('扫描目录：');
+        $data = $this->scanner->scanDir(Env::path_framework_generated_code);
+        if ($data) {
+            $this->printer->warning($i . '、generated生成目录代码code清理...', '系统');
+            $this->system->exec('rm -rf ' . Env::path_framework_generated_code);
+        }
         $i += 1;
         // 扫描代码
         $this->printer->note($i . '、清理模板缓存', '系统');
@@ -99,12 +107,12 @@ class Upgrade implements \Weline\Framework\Console\CommandInterface
                 require $register;
             }
         }
-        // 过滤可以用的modules
-        $modules = Env::getInstance()->getModuleList(true);
+        // 重新可以用的modules
+        $modules = Env::getInstance()->getActiveModules();
         /**@var Sort $sort */
         $sort                = ObjectManager::getInstance(Sort::class);
         $module_dependencies = $sort->dependenciesSort($modules);
-        Env::write(Env::path_MODULE_DEPENDENCIES_FILE, '<?php return ' . var_export($module_dependencies, true) . ';?>');
+        Env::write(Env::path_MODULE_DEPENDENCIES_FILE, '<?php return ' . w_var_export($module_dependencies, true) . ';?>');
         /**@var Handle $module_handle*/
         $module_handle = ObjectManager::getInstance(Handle::class);
         // 安装Setup信息
